requirements.txt
asn1crypto==1.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cowsay==2.0.3
cryptography==2.7
Django==2.2.10
django-appconf==1.0.3
django-imagekit==4.0.2
-e git+git://github.com/telebotter/django-telegrambot.git#egg=django-telegrambot
python-telegram-handler==2.2
emoji==0.5.4
future==0.17.1
idna==2.8
mysqlclient==1.4.4
openmensa-api==1.1
pilkit==2.0
Pillow==6.2.0
pycparser==2.19
python-telegram-bot==12.1.1
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.6
django-markdownx
django-telegram-login
pyyaml>=5.3.1
